+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Stop profiler when app moves to background (#2331)
 - Clean up old envelopes (#2322)
+- Crash when starting a profile from a non-main thread (#2345)
 - SentryCrash writing nan for invalid number (#2348)
 
 ## 7.29.0